requirements.txt
# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0

# Machine Learning
tensorflow==2.15.0
numpy==1.23.5
scikit-learn==1.3.2

# NLP
nltk==3.8.1

# SharePoint
Office365-REST-Python-Client==2.5.3

# Utilitaires
python-dotenv==1.0.0
requests==2.31.0

# Development (optionnel)
autopep8==2.0.4
pylint==3.0.3
pytest==8.3.5
